# Fix data race when running `go-git` tests